remove target=“_blank”
Let users have the freedom to choose how they want to open links. Our audience is tech-savvy. If they want to open a link in a new tab or window, they know how to do it.

Moreover, it is really convinient to remove target=_blank to use Micronaut documentation as a [Safari Web App](https://support.apple.com/en-us/104996)